"Рефакторинг тестов стратегий: добавлена общая функция `generalCardTest` для проверки карточек, актуализированы данные тестов."
