+++ src/strategies/SamokatStrategy.ts
@@ -1,6 +1,6 @@
 import { ParserStrategy } from "@/core/ParserStrategy";
 import { getUnitParsedWeight, roundNumber } from "@/utils/converters";
-import { UnitLabel } from "@/types/IStrategy";
+import type { UnitLabel } from "@/types/IStrategy";
 
 export class SamokatStrategy extends ParserStrategy {
   constructor() {
@@ -16,9 +16,9 @@ export class SamokatStrategy extends ParserStrategy {
   }
 
   parsePrice(cardEl: HTMLElement): number {
-    const priceString = cardEl.querySelector(this.selectors.price)?.textContent;
+    const priceString = this.selectors?.price ? cardEl.querySelector(this.selectors.price)?.textContent || "" : "";
     console.log("parsed price text", priceString);
-    const num = priceString?.replace(/[^\d,\.]/g, "").replace(",", ".") ?? "";
+    const num = priceString?.replace(/[^\d,.]/g, "").replace(",", ".") ?? "";
     const v = parseFloat(num);
     if (isNaN(v)) throw new Error("Цена не распознана: " + priceString);
     return v;
