+++ src/strategies/__test_data__/samokat/card1.json
@@ -1,0 +1,20 @@
+{
+  "html": "<a href=\"/product/4d75b935-5a26-11ea-a7b3-0050560306e1\"><div class=\"ProductCard_root__OCLMl\"><div class=\"ProductCardImage_root___xdPV\"><img alt=\"Молоко Домик в деревне, 3,2%, ультрапастеризованное, 925 мл\" loading=\"lazy\" decoding=\"async\" data-nimg=\"fill\" style=\"position: absolute; height: 100%; width: 100%; inset: 0px; color: transparent;\" src=\"https://cm.samokat.ru/processed/m/product_card/ff5e3b0f-3bdf-4584-9fad-c3fbb5c7268d.jpg\"><div style=\"transform: translateY(0%);\" class=\"ProductCardOverlay_root__z6uaf ProductCardOverlay_root_hidden___YjHT\"><div class=\"ProductCardOverlay_slide__LnQ0b\"><span style=\"color: rgb(255, 255, 255);\" class=\"Text_text__7SbT7 Text_text--type_h1Bold__SwxtG\">1</span></div><div class=\"ProductCardOverlay_slide__LnQ0b\"><span style=\"color: rgb(255, 255, 255);\" class=\"Text_text__7SbT7 Text_text--type_h1Bold__SwxtG\">2</span></div><div class=\"ProductCardOverlay_slide__LnQ0b\"><span style=\"color: rgb(255, 255, 255);\" class=\"Text_text__7SbT7 Text_text--type_h1Bold__SwxtG\">3</span></div><div class=\"ProductCardOverlay_slide__LnQ0b\"><span style=\"color: rgb(255, 255, 255);\" class=\"Text_text__7SbT7 Text_text--type_h1Bold__SwxtG\">4</span></div><div class=\"ProductCardOverlay_slide__LnQ0b\"><span style=\"color: rgb(255, 255, 255);\" class=\"Text_text__7SbT7 Text_text--type_h1Bold__SwxtG\">5</span></div><div class=\"ProductCardOverlay_slide__LnQ0b\"><span style=\"color: rgb(255, 255, 255);\" class=\"Text_text__7SbT7 Text_text--type_h1Bold__SwxtG\">6</span></div><div class=\"ProductCardOverlay_slide__LnQ0b\"><span style=\"color: rgb(255, 255, 255);\" class=\"Text_text__7SbT7 Text_text--type_h1Bold__SwxtG\">7</span></div><div class=\"ProductCardOverlay_slide__LnQ0b\"><span style=\"color: rgb(255, 255, 255);\" class=\"Text_text__7SbT7 Text_text--type_h1Bold__SwxtG\">8</span></div><div class=\"ProductCardOverlay_slide__LnQ0b\"><span style=\"color: rgb(255, 255, 255);\" class=\"Text_text__7SbT7 Text_text--type_h1Bold__SwxtG\">9</span></div><div class=\"ProductCardOverlay_slide__LnQ0b\"><span style=\"color: rgb(255, 255, 255);\" class=\"Text_text__7SbT7 Text_text--type_h1Bold__SwxtG\">10</span><span style=\"color: rgb(255, 255, 255);\" class=\"Text_text__7SbT7 Text_text--type_h4Bold__2MTmo ProductCardOverlay_subtitle__41TER\">Больше нет</span></div></div></div><div class=\"ProductCard_content__EjT48\"><div class=\"ProductCard_details__S6PcT\"><div class=\"ProductCard_name__2VDcL\" title=\"Молоко Домик в деревне, 3,2%, ультрапастеризованное, 925 мл\"><span style=\"color: rgb(89, 89, 89);\" class=\"Text_text__7SbT7 Text_text--type_p3SemiBold__nZftu\">Молоко Домик в деревне, 3,2%, ультрапастеризованное</span></div><div class=\"ProductCard_specification__Y0xA6\" style=\"color: rgb(0, 183, 73);\"><span style=\"color: rgb(166, 166, 166);\" class=\"Text_text__7SbT7 Text_text--type_p3SemiBold__nZftu\">925&nbsp;мл</span><span style=\"color: rgb(166, 166, 166);\" class=\"Text_text__7SbT7 Text_text--type_p3SemiBold__nZftu ProductCard_separator__8t6DW\">·</span><span style=\"color: rgb(0, 183, 73);\" class=\"Text_text__7SbT7 Text_text--type_p3SemiBold__nZftu\">Теперь дешевле</span></div></div><div class=\"ProductCard_actions__76n23\"><div class=\"ProductCardActions_root__Jcqm6 ProductCardActions_root_size_small__Tz_eu\"><div class=\"ProductCardActions_text__3Uohy\"><span style=\"color: rgb(64, 64, 64);\" class=\"Text_text__7SbT7 Text_text--type_p2SemiBold__JDK1U\"><span>149&nbsp;₽</span></span></div><div class=\"ProductCardActions_action__m4aYK ProductCardActions_increase__Od4HR\"><div class=\"Icon_icon__WhKMY\" style=\"height: 24px;\" aria-hidden=\"true\"><svg xmlns=\"http://www.w3.org/2000/svg\" width=\"24\" height=\"24\" fill=\"none\"><path fill=\"currentColor\" fill-rule=\"evenodd\" d=\"M11.063 2.938a.937.937 0 1 1 1.874 0v6.5c0 .897.728 1.624 1.626 1.624h6.5a.937.937 0 1 1 0 1.876h-6.5c-.898 0-1.626.727-1.626 1.624v6.501a.937.937 0 1 1-1.874 0v-6.5c0-.898-.728-1.626-1.626-1.626H2.938a.937.937 0 1 1 0-1.874h6.5c.897-.001 1.624-.728 1.624-1.626V2.938\" clip-rule=\"evenodd\"></path></svg></div></div></div></div></div></div></a>",
+  "expectedParsedPrice": 149,
+  "expectedParsedQuantity": {
+    "unitLabel": "1 л",
+    "multiplier": 1.081
+  },
+  "expectedUnitPrice": {
+    "price": 161,
+    "label": "1 л",
+    "styles": {
+      "display": "inline-block",
+      "color": "rgb(0, 0, 0)",
+      "padding": "2px 6px",
+      "borderRadius": "4px",
+      "fontWeight": "600",
+      "fontSize": "15px"
+    }
+  }
+}
