+++ src/strategies/__tests__/SamokatStrategy.test.ts
@@ -1,11 +1,11 @@
 import { describe } from "vitest";
-import { AuchanStrategy } from "@/strategies";
+import { SamokatStrategy } from "@/strategies";
 import { loadAllTestCards } from "../__test_data__/loadTestData";
 import { generalCardTest } from "@/strategies/__tests__/generalTest";
 
-describe("AuchanStrategy", () => {
-  const strategy = new AuchanStrategy();
-  const testCards = loadAllTestCards("auchan");
+describe("SamokatStrategy", () => {
+  const strategy = new SamokatStrategy();
+  const testCards = loadAllTestCards("samokat");
 
   const currentCardTest = generalCardTest(strategy);
   testCards.forEach(currentCardTest);
